LayoutTests/fast/dom/Node/nullable-parameters-expected.txt

Tests the behavior of a few Node API parameters that should be nullable

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS xmlDoc.documentElement.lookupPrefix('undefined') is "c"
PASS xmlDoc.documentElement.lookupPrefix(null) is null
PASS xmlDoc.documentElement.lookupPrefix(undefined) is null

PASS xmlDoc.documentElement.lookupNamespaceURI('undefined') is "test"
PASS xmlDoc.documentElement.lookupNamespaceURI(null) is null
PASS xmlDoc.documentElement.lookupNamespaceURI(undefined) is null

PASS element.isDefaultNamespace('undefined') is true
PASS element.isDefaultNamespace(undefined) is false
PASS element.isDefaultNamespace('null') is true
PASS element.isDefaultNamespace(null) is false
PASS successfullyParsed is true

TEST COMPLETE
